Update wing1-spline.py to output STL directly
Adam asked me to output the stl files directly in this post:
  https://groups.google.com/g/cadquery/c/DcLv4ZD3r5Q/m/Nv0z2u0AAwAJ

So I've added the code to output directly from my code, here is the change
to wing1-spling.py

    67	result = cq.Workplane("XY").spline(fMh49).close().extrude(span)
    68	#pprint(vars(result))
    69	import io
    70	f = io.open('wing1-spline-direct-0.5.stl', 'w+')
    71	cq.exporters.exportShape(result, cq.exporters.ExportTypes.STL, f, 0.5)
    72	f.close()

And here is the output. The first 4 files were output by cq-editor
and they are quite different from any of the versions that
wing1-spline.py output:

 (cqgui-master) wink@3900x:~/prgs/CadQuery/projects/wing1 (master)
 $ wc -l *.stl
     926 mh49_50x50_spline_stl_precision_0.01.stl
     926 mh49_50x50_spline_stl_precision_0.1.stl
    1654 mh49_50x50.stl
    1682 mh49_50x50_te0.5mm.stl
    3110 wing1-spline-direct-0.001.stl
    2942 wing1-spline-direct-0.01.stl
    2690 wing1-spline-direct-0.1.stl
  103000 wing1-spline-direct-0.5.stl
  116930 total

Looking at cadquery/occ_impl/shapes.py it seems there maybe a problem withcalls to exportStl as its parameters are:

    def exportStl(self, fileName, precision=1e-3, angularPrecision=0.1):

Where as the call to exportStl from cadquery/occ_impl/exporters.py is only passing2 which means the angularPrecision parameter is the default value, 0.1:

        elif exportType == ExportTypes.STL:
            shape.exportStl(outFileName, tolerance)

Which in and of itself isn't a problem, But the tolerance parameter
in exportShape:

    def exportShape(shape, exportType, fileLike, tolerance=0.1):

Has a default value the same as the angularPrecision but it's being
passed as the precision which defaults to 1e-3 (0.001). This smells
funny.